Support QingStor CLI
QingStor CLI is a client tool for user to communicate with QingStor
server. As an independent service, we add new service 'qs', and use
the term as the subcommand which is different from 'iaas' service.

To see all the QingStor service actions and their usages, use command
'qingcloud qs -h' and 'qingcloud qs <action> -h'.

This change depends on the library qingcloud-sdk-python >= 1.0.0